[BCO-8378] Support higher values for VoucherConstraintsApplications.count, change type Integer -> Long

+++ src/main/java/com/scayle/adminapi/model/VoucherConstraintsApplications.java
@@ -15,7 +15,7 @@ public class VoucherConstraintsApplications implements ApiObjectInterface {
     * 
     */
     @SerializedName("count")
-    Integer count;
+    Long count;
 
     /**
     *
